test(message-list): add tests for message active/open/toggle side effects
