+++ feature/mail/message/list/internal/src/test/kotlin/net/thunderbird/feature/mail/message/list/internal/ui/state/sideeffect/BaseSideEffectHandlerTest.kt
@@ -2,6 +2,7 @@ package net.thunderbird.feature.mail.message.list.internal.ui.state.sideeffect
 
 import androidx.compose.ui.graphics.Color
 import kotlinx.collections.immutable.ImmutableSet
+import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentMapOf
 import kotlinx.collections.immutable.persistentSetOf
 import net.thunderbird.core.common.action.SwipeAction
@@ -16,7 +17,9 @@ import net.thunderbird.feature.mail.message.list.domain.model.SortType
 import net.thunderbird.feature.mail.message.list.preferences.ActionRequiringUserConfirmation
 import net.thunderbird.feature.mail.message.list.preferences.MessageListPreferences
 import net.thunderbird.feature.mail.message.list.ui.state.Account
+import net.thunderbird.feature.mail.message.list.ui.state.ComposedAddressUi
 import net.thunderbird.feature.mail.message.list.ui.state.Folder
+import net.thunderbird.feature.mail.message.list.ui.state.MessageItemUi
 import net.thunderbird.feature.mail.message.list.ui.state.MessageListMetadata
 import net.thunderbird.feature.mail.message.list.ui.state.MessageListState
 
@@ -25,6 +28,30 @@ open class BaseSideEffectHandlerTest {
     protected fun createReadyWarmingUpState() = MessageListState.WarmingUp(
         metadata = createReadyMetadata(),
         preferences = createMessageListPreferences(),
+    )
+
+    protected fun createLoadedMessagesState(
+        activeMessage: MessageItemUi? = null,
+    ) = MessageListState.LoadedMessages(
+        metadata = createReadyMetadata().copy(activeMessage = activeMessage),
+        preferences = createMessageListPreferences(),
+        messages = persistentListOf(),
+    )
+
+    protected fun createSelectingMessagesState() = MessageListState.SelectingMessages(
+        metadata = createReadyMetadata(),
+        preferences = createMessageListPreferences(),
+        messages = persistentListOf(),
+    )
+
+    protected fun createLoadingMessagesState(
+        progress: Float = 0f,
+        metadata: MessageListMetadata = createMetadata(),
+        preferences: MessageListPreferences = createMessageListPreferences(),
+    ) = MessageListState.LoadingMessages(
+        progress = progress,
+        metadata = metadata,
+        preferences = preferences,
     )
 
     protected fun createReadyMetadata() = MessageListMetadata(
@@ -77,5 +104,24 @@ open class BaseSideEffectHandlerTest {
         sortCriteriaPerAccount = persistentMapOf(),
         activeMessage = null,
         isActive = true,
+    )
+
+    fun createMessageItemUi(
+        id: String = "1",
+    ) = MessageItemUi(
+        state = MessageItemUi.State.Unread,
+        id = id,
+        messageReference = "ref-$id",
+        account = Account(id = AccountIdFactory.create(), color = Color.Unspecified),
+        senders = ComposedAddressUi(displayName = "sender"),
+        subject = "subject",
+        excerpt = "excerpt",
+        formattedReceivedAt = "Jan 2026",
+        hasAttachments = false,
+        starred = false,
+        encrypted = false,
+        answered = false,
+        forwarded = false,
+        selected = false,
     )
 }
